Added more tests, applied Clippy, and separated grid into spatial.rs and pointer.rs.
