+++ src/molecule.rs
@@ -6,7 +6,7 @@ use std::fmt::{Display, Formatter};
 use ruscii::spatial::{Direction, Vec2};
 use ruscii::terminal::Color;
 use ruscii::terminal::Color::{LightGrey, Red, White};
-use crate::grid::Cellular;
+use crate::spatial::Cellular;
 use crate::molecule::Element::{C, H, O};
 use crate::molecule::BondOrder::{Double, Single, Triple};
 use crate::molecule::BondOrientation::{Horiz, Vert};
@@ -75,11 +75,7 @@ impl Cell {
     }
 
     pub(crate) fn is_atom(&self) -> bool {
-        if let Cell::Atom(_) = self {
-            true
-        } else {
-            false
-        }
+        matches!(self, Cell::Atom(_))
     }
 }
 
@@ -193,7 +189,7 @@ impl BondOrientation {
 
 impl Clone for Bond {
     fn clone(&self) -> Bond {
-        Bond { pos: self.pos.clone(), order: self.order.clone(), orient: self.orient.clone() }
+        Bond { pos: self.pos, order: self.order, orient: self.orient }
     }
 }
